requirements.txt
prefect>=2.18
polars>=0.20
requests>=2.31
python-dotenv>=1.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
pydantic>=2.6
pandera>=0.18
pyarrow>=15.0
